Foundations for new stylesheet
Most of the front page has been successfully revamped. There is still more
to do with internal moderator pages, the submission page, etc., but the
foundations are there.

A set of six flair icons is now supported for easy identification of
thread categories and uncategorized threads. Some of the new CSS changes,
for example the visual divisions between content, sidebar, and
announcement bar, have been largely inspired by /r/boxed's layout.

Night mode is currently only a work in progress. The division between day
and night mode is present in the code, but night mode's color scheme has
not yet been decided on.